backup_database/requirements.txt
# thư viện cần thiết
python-dotenv
schedule